Remove the db_storage_v0 file.
